requirements.txt
alembic==1.4.1
azure-core==1.8.2
azure-storage-blob==12.5.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
cryptography==3.2.1
cycler==0.10.0
databricks-cli==0.13.0
docker==4.3.1
entrypoints==0.3
Flask==1.1.2
gitdb==4.0.5
GitPython==3.1.11
gorilla==0.3.0
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.0
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mlflow==1.11.0
msrest==0.6.19
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.4
parameterized==0.7.4
Pillow==8.0.1
prometheus-client==0.8.0
prometheus-flask-exporter==0.18.1
protobuf==3.13.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
querystring-parser==1.2.4
requests==2.24.0
requests-oauthlib==1.3.0
scipy==1.5.3
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.13
sqlparse==0.4.1
tabulate==0.8.7
tqdm==4.51.0
urllib3==1.25.11
websocket-client==0.57.0
Werkzeug==1.0.1
